red(sampling): Fix sampling modal race condition
