+++ static/app/types/sampling.tsx
@@ -172,11 +172,6 @@ export type SamplingRule = {
    * Indicates if the rule is enabled for server-side sampling
    */
   active?: boolean;
-  /**
-   * A rule without a condition (Else case) always have to be 'pinned'
-   * to the bottom of the list and cannot be sorted.
-   */
-  bottomPinned?: boolean;
 };
 
 export type SamplingDistribution = {
